Modularize package.py
This has multiple thousand lines of code. we can split each
container in its own package.
