+++ .gitignore
@@ -1,3 +1,4 @@
 /build/
 /dist/
 /test-build.env
+**/__pycache__/
